common * 実装見直し。

+++ common/src/test/java/cherry/spring/common/lib/paginate/EachSideStrategyTest.java
@@ -66,9 +66,9 @@ public class EachSideStrategyTest {
 		check(new Range(7, 8), strategy.calculate(9, 10));
 	}
 
-	private void check(Iterable<Integer> a, Iterable<Integer> b) {
-		Iterator<Integer> ia = a.iterator();
-		Iterator<Integer> ib = b.iterator();
+	private void check(Iterable<Long> a, Iterable<Long> b) {
+		Iterator<Long> ia = a.iterator();
+		Iterator<Long> ib = b.iterator();
 		while (ia.hasNext() && ib.hasNext()) {
 			assertEquals(ia.next(), ib.next());
 		}
